// src/app/admin/display-decorators/models/property-types.enum.ts
/**
 * @description Property types that can be handles by the display decorator
 */
export enum PropertyTypes {
    class = 'class',
    array = 'array',
    arrayTable = 'array-table',
    input = 'input',
    checkbox = 'checkbox',
    select = 'select',
    searchSelect = 'searchSelect',
    none = 'none',
}